Use parallel socket to transfer file descriptors
For gRPC/Cap'n Proto, rather than relying on inheritance of FDs
between parent and child, instead keep the bootstrap socket open
and send FDs over it in parallel with the main RPC message.

Include a random number as a nonce in the sendmsg to ensure that
the file descriptor matches up.